django/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
numpy==1.26.4
pandas==2.2.2
aiohttp==3.9.5
uvicorn==0.30.0
uvloop==0.19.0
httptools==0.6.1
psycopg[binary]==3.1.19
django-cors-headers==4.3.1
asgiref==3.8.1
